Correctly type empty variables in TypeScript (#4404)
Summary:
Previously the compiler produced `{}` for empty objects in TypeScript. This is for example the case for the `...$variables` type, if no variables are present.

The problem is that `{}` does not mean "empty object" in TypeScript. It means "any value except null and undefined".
Because of this you're able to input **anything** besides those two values as variables for an operation without variables.

I opted to produce the more appropriate `Record<PropertyKey, never>`, which effectively represents an object without properties.

Pull Request resolved: https://github.com/facebook/relay/pull/4404

Reviewed By: alunyov

Differential Revision: D48989147

Pulled By: captbaritone

fbshipit-source-id: 67bad8a670496a3caf7b2a3c439a8ff0b283793d