Fix type generation of nullable types in TypeScript (#4380)
Summary:
Closes: https://github.com/facebook/relay/issues/4379

## Release Notes

Previously, the relay-compiler would generate a union of the actual type and `null` (`T | null`) for nullable fields when targeting TypeScript. This was not correct, since the Relay store can also return `undefined` in the case of missing data. This version now produces a union of the actual type, `null` and `undefined` (`T | null | undefined`).
Consequently you now also have to check both the `null` and `undefined` case, before accessing a field's value:

```ts
const data = useLazyLoadQuery(/* ... */);

// Notice the `!=` instead of `!==` to check against null and undefined
if (data.nullableField != null)
{
  // Safely access data.nullableField
}
```

Since this is a pretty big change, we're offering the `typescriptExcludeUndefinedFromNullableUnion` feature flag in the `relay-compiler` config to keep the old type generation behavior for now:

```json
{
  "language": "typescript",
  "typescriptExcludeUndefinedFromNullableUnion": true
}
```

Just be aware that we will remove this feature flag in a future release and that you might run into an unexpected `undefined`, if you're not checking for `undefined` before accessing a field - especially if it's a client-only field.

Pull Request resolved: https://github.com/facebook/relay/pull/4380

Reviewed By: alunyov

Differential Revision: D48989476

Pulled By: captbaritone

fbshipit-source-id: 52c5f215298de50e8506245a9d6696e1e935fc2a