Improve inline fragment suggestions for abstract types (#4453)
Summary:
Previously the LSP would suggest the abstract type itself (why?) and either the members of a union or the object types implementing an interface when writing an inline fragment for an abstract type.

I've removed the suggestion of the abstract type itself and am instead including all the interfaces that are implemented by the object types that are primarily suggested. This is great for patterns like the Stage 6a mutation error pattern, where you have a union where each member implicitly implements a shared interface. With this change this shared interface would be suggested by the LSP as well, instead of just the concrete "errors".

Pull Request resolved: https://github.com/facebook/relay/pull/4453

Reviewed By: captbaritone

Differential Revision: D49956467

Pulled By: alunyov

fbshipit-source-id: db4581532f2dd74d1262eb8f9a0c4952480ccee8