Allow configurable `@defer`/`@stream` names (#4467)
Summary:
This is not an attempt to get Relay to conform to the current defer/stream spec draft, but rather just a small step in that direction by allowing configurable directive & argument names.

Currently Relay requires the argument `initial_count` on `stream`, but the GraphQL spec uses camelCase for all names.

This PR updates Relay compiler to follow the same pattern used by `ConnectionInterface`, introducing `DeferStreamInterface` to allow configuring the names and arguments of the defer & stream directive.

The first commit adds the configuration updating all touch points to reference the configuration.

The second commit changes the defaults to camelCase, to match the defaults used in `ConnectionInterface`. This makes the diff substantially larger as many tests and fixtures need to be updated. I'm happy to remove this commit from this PR if desired.

Pull Request resolved: https://github.com/facebook/relay/pull/4467

Reviewed By: alunyov

Differential Revision: D50043794

Pulled By: captbaritone

fbshipit-source-id: e27f159ef8e22dd197b85e2653b9541c66594df1