Add areEqualOwners to check for structural equality of fragment owners. (#4500)
Summary:
There is a bug in the experimental hooks implementation.

useRefetchableFragment will return a new data object every time the refetch function is called, even if the data has not changed. This is because refetch creates a new fragment owner with createOperationDescriptor and refetch variables, and passes it to useFragmentInternal as part of the fragment key.

Later in useFragmentInternal, we call areEqualSelectors to check if the fragment selector has changed. The owner is part of the selector, but we only check if the owner is the same object. However, it's possible that the owner is the same but just a new instance of the same object.

Proposed fix:
* Add areEqualOwners to check that owners are equivalent. This would prevent an endless loop of refetch/setState between components using useRefetchableFragment and the refetch call in the useEffect.

Pull Request resolved: https://github.com/facebook/relay/pull/4500

Test Plan:
* New unit test-case to useRefetchableFragmentNode-test.
* Updated tests for RelayModernSelector.

Reviewed By: voideanvalue

Differential Revision: D50524883

Pulled By: alunyov

fbshipit-source-id: 6d9df3abc80ab997a59bf961e7a721da6ad769e3