rename relay resolver test to be more specific
Summary:
## Context

When compiling relay resolvers we are missing validation on undefined global variables within resolvers. The global variables can be passed via resolver field or fragment arguments.

We want to add a suite of regression tests for the following cases:

- Relay resolver yields a scalar field without an associated fragment e.g. live or derived resolvers
- Relay resolver yields a scalar field with an associated fragment
  - Passing both field and fragment arguments
  - Fragment args only
   - Field args only
  - Multiple undefined global variables
- Linked field aka client edges

## This diff

We have existing tests for scalar field with a fragment + field args only. Renaming it to be a bit more specific

Rename test on undefined global variables validation in relay resolvers to be more specific to undefined field arguments

Reviewed By: captbaritone, tyao1

Differential Revision: D50615718

fbshipit-source-id: 22b7e526ae2dbbc44e05e7ffbd3241b557da9964