add test that field args are dropped from scalar relay resolvers during transform
Summary:
## Context

When compiling relay resolvers we are missing validation on undefined global variables within resolvers. The global variables can be passed via resolver field or fragment arguments.

We want to check whether the relay_resolvers transform is dropping field arguments for relay resolvers without a fragment attached + returning a scalar value.

## This diff

Adds a test on the relay_resolvers transform to verify the above.

Add a test that field arguments are dropped from scalar relay resolvers during the `relay_resolvers` transform step and added to the `RelayResolverMetadata`.

Reviewed By: voideanvalue

Differential Revision: D50624879

fbshipit-source-id: bdce4180f79b409271100d5810bcc7b0c8523bc2