rephrase "is done"
it says "when the mutation is done", I had to read it several times and finally relied on the diagram below.

This "is done" actually means "succeeded or failed". That may be a bit different from our daily language: if I say the task is done, I usually don't mean, "the task is done. It failed."

It usually means "the task is done. It is completed. (and naturally I mean succeeded)."

So here, to let the reader understand it by reading it once, it can be rephrased to "when the mutation is done (either success or failure)", or "when the mutation has succeeded or failed"