fix RelayResolverModel-test
Summary: This test does not actually add the todo object to the `TODO_STORE` that it reads (and expects to exist), but the current implementation of `RelayReader` creates the todo record regardless so the test is passing. After the fix for T162471299 lands, we need the changes in this diff so that the test passes.

Reviewed By: voideanvalue

Differential Revision: D51436475

fbshipit-source-id: ac4ff215c9d2b7e14b2ac2359dbe038da7c20c87