implement fmt::Display for SchemaChange
Summary:
## Context

During the investigation of S380145, it was useful to check whether the compiler build time was regressing due to a full build being triggered.

We'd like to log some data on whether a build was incremental or not and if not, why. If we see that we are often falling back to a full builds due to schema changes, we can revisit the classification of schema changes that necessitate a full build.

## This diff

As titled - to log the unsafe schema change that requires a full build

Reviewed By: tyao1

Differential Revision: D51515403

fbshipit-source-id: 9bdc1d8070fd3b0aff6b930a8fd07c61a1826af8