add helpers to log bool
Summary:
## Context

During the investigation of S380145, it was useful to check whether the compiler build time was regressing due to a full build being triggered. We'd like to log some data on whether a build was incremental or not and if not, why. If we see that we are often falling back to a full builds due to schema changes, we can revisit the classification of schema changes that necessitate a full build.

## This diff

Adds some helpers to allow logging boolean values directly + abstract away the string casting (Scuba doesn't actually have a boolean data type, and we are already logging other booleans as normals: https://www.internalfb.com/intern/wiki/Scuba/Quick_Start_Guide/Data_Types/).

Reviewed By: tyao1

Differential Revision: D51515404

fbshipit-source-id: e80b2324279ff101c22f894a31620cc898985283