log incremental build details during build_project
Summary:
## Context

During the investigation of S380145, it was useful to check whether the compiler build time was regressing due to a full build being triggered. We'd like to log some data on whether a build was incremental or not and if not, why. If we see that we are often falling back to a full builds due to schema changes, we can revisit the classification of schema changes that necessitate a full build.

## This diff

Adds the incremental build logging. Splits some functions to make that easier.
Fields added-
* `is_incremental_build`
* `has_breaking_schema_changes`
* `schema_change` if the breaking schema change was not due to docblock/extension

Reviewed By: tyao1

Differential Revision: D51515402

fbshipit-source-id: a977ca81bc8bf94a0f741083feac0c886c1da1a0