useFragmentInternal should suspend if environment changes and there is a query in flight
Summary:
There is a bug in the new hooks implementation, where the `useFragmentInternal` would not suspend after the change of the environment (while the fragment selector are the same) with the active operation in flight.

This diff fixes that by checking if the environment has changed and we would throw the promise for pending operation in flight.

Reviewed By: voideanvalue

Differential Revision: D52516772

fbshipit-source-id: 13a61114e6b7718af548cf45d2e4685bcd121993