fix partition for conditions and inline fragments without types
Summary:
After this change, conditions and inline fragments will be partitioned into "copy" or "keep" based on whether any of their children are "copy" since I believe they are intended to be moved as a single chunk.

This addresses a bug where inline fragments without types and conditions are always copied to inline fragments on concrete types, even if their selection fields are all defined in client schema extensions/ on the server (not using resolvers).

Reviewed By: captbaritone

Differential Revision: D53793622

fbshipit-source-id: ed97f0f037bcaa08e84c43218d11f01399f86915