add resolver metadata to fixture test output
Summary: This change adds a second pass for the `relay_resolvers` transform after the `relay_resolvers_abstract_types` transform to the fixture transform for the abstract types test so that we can examine the resolver metadata in the fixture test output. This tells us, in the case that some of the concrete types' field implementations are resolver fields while the interface's are not, whether the selections copied to inline fragments on concrete types reflect the correct resolver field implementation on the concrete type.

Reviewed By: captbaritone

Differential Revision: D53789647

fbshipit-source-id: c94dbbc3e10317026a8040a66ff59a6a3e8dccc9