cache whether interface's implementing objects are all server-defined
Summary:
If all of an interface's implementing objects are server-defined, then it's not possible that one of the implemented fields are client-defined (and need to be duplicated to inline fragments on the respective concrete type).

In that case, we can skip transforming the selections entirely.

This diff adds support for caching whether the interface's implementing objects are server-defined and if so, skips the transform in `transform_selections_given_parent_type` by returning `TransformedValue::Keep`.

Reviewed By: captbaritone

Differential Revision: D54390572

fbshipit-source-id: e15a877df5e19e6585353cffc2961517d11f83c8