refactor client edge model resolver field type to an object map of model resolvers
Summary:
## Context

In M2 we want to support defining a resolver that returns an interface, implemented by strong concrete types. The `modelResolver` field on a `ClientEdgeToClientObject` type is used in `RelayReader` when reading a client edge to verify that the object exists before reading its child fields.

Previously, `ClientEdgeToClientObject` reader nodes were structured as:
```
{
   ...
   'modelResolver': {..}
   'backingField': {..}
   'linkedField': {..}
   ...
}
```
## This diff
This diff refactors the structure of `ClientEdgeToClientObject` nodes to:
```
{
   ...
   'modelResolvers': {
      'ConcreteTypeA': {..}
      'ConcreteTypeB': {..}
   }
   'backingField': {..}
   'linkedField': {..}
   ...
}
```

where the type of the `modelResolvers` field in runtime code is `{[string]: RelayResolver | RelayLiveResolver}`.

There will only be one concrete type if the client edge points to a concrete object, i.e.
```
{
   ...
   'modelResolvers': {
      'ConcreteObjectType': {..}
   }
   'backingField': {..}
   'linkedField': {..}
   ...
}
```

In this diff we are only handling client edges that point to a concrete object.
Hence, the model resolvers for client edges pointing to an interface are hardcoded to an empty object.

Note: model resolvers for `ClientEdgeToClientUnion` nodes are not yet generated here.

Reviewed By: captbaritone

Differential Revision: D54104542

fbshipit-source-id: d20374211503015ca746be17bf29de3da9d38a27