Pass schema changes to IR dependency analyzer
Summary:
This diff prepares for incrementally building schema changes by converting the boolean representing "breaking schema changes"/"incremental builds" to an enum that represents "no build", "full build", and "incremental build".

The changes begin in `check.rs` and most of the rest of the code is piping the new enums through the previous code pathways. I convert the new enum back into a boolean before passing it to `get_reachable_ir`. A stacked diff will handle the actual incremental changes in that function.

Reviewed By: tyao1

Differential Revision: D54547132

fbshipit-source-id: 53e92bfa453999d3cc3aa1be4216519e946c7d44