update GraphQL parser to consume escaped characters
Summary:
### Context

We got [feedback](https://fb.workplace.com/groups/relay.support/permalink/25937328725889064/) in Relay Support about some generated artifacts being deleted in D54762067 (when `parseData` was declared in a variable outside the component specifically). It turns out that the RegEx syntax added in the diff was not being parsed correctly by the GraphQL extractor. So, the GraphQL source text was not extracted from the file and the generated artifact was removed.

## This diff
Adds logic to consume escaped characters. Note that comments and strings are already consumed generally.

I also changed the iterator to be peekable. The example application for this diff is in the regex `/text\/*/`: without peeking, `consume_identifier` consumes `text\`  though `\` should be used as an escape.

It seems we've run into similar issues before which led us to consider switching to a stricter GraphQL parser (see e.g. D21256605 from 2020). Maybe we should revisit that idea?

Reviewed By: evanyeung

Differential Revision: D54879768

fbshipit-source-id: cccb3311acf7188b642c478f2a68f18d8d7eb66b