push validation on weak type/legacy syntax to before model resolver creation
Summary:
## Context

If the model that the client edge points to is a weak type or if the client edge is pointing to an object defined with legacy syntax, we skip creating model resolvers (for weak client edges, the model resolvers are unnecessary and for legacy edges, the model resolvers don't work properly since the generated id fragment paths don't match).

Before, we were checking these two properties in `build_ast`, but we can do so in the `client_edges` transform to avoid generating the model resolver IR when unnecessary.

## This diff

Refactors the existing implementation to check for weak and legacy types in `client_edges` transform.

Reviewed By: captbaritone

Differential Revision: D54556696

fbshipit-source-id: 3e4444856bc8ac66e24bfc12ddcd19745d4c019c