Incrementally build Object schema changes
Summary:
This diff follows the pattern of doing an incremental change for enums but does so for object changes.

This diff adds incremental builds for changes to objects in the schema. It has to cover some special cases related to objects: interfaces and unions. The union tracking is fairly simple, if an object changes and that object is in a union, the union is added to the set of changes.

Interfaces are a bit more compex as it depends on if an object has an id or not. If an object has an id and an interface, changes to the id or interfaces will trigger changes on all interfaces for that object. This is due to the interface id transform.

Changes were also made to the dependency analyzer to cover places that union and interface types can show up beyond linked fields: fragments and inline fragments.

Reviewed By: tyao1

Differential Revision: D54604246

fbshipit-source-id: 9a8a3076f529cf04e22b7867a0ed1d4b6d646b9e