Crate for extracting type information from JS definitions
Summary:
Add a crate to extract flow type information from JS definitions

This diff adds a schema-extractor for extracting type information from JS definitions that can be used for constructing GraphQL types.
The general idea for the crate is that it should be able to be used to extract information for RelayResolver flavor or general gql flavor.
Next we will build another crate to consume this information and resolve these raw type identifiers to their RelayResolver definitions

Reviewed By: captbaritone

Differential Revision: D55044177

fbshipit-source-id: c79582b478d4579bbdd64b03e8dbeda7a1ec92cd