Prevent name collision between derived refetchable query and fragment name (#4663)
Summary:
As identified in https://github.com/facebook/relay/issues/4654 it's possible to use an existing fragment name as the `queryName` argument of `refetchable` and end up with a collision. The result is that we end up importing the fragment's artifact when we expect to import the queries. It's possible we even overwrite that same path with two artifacts.

~One thing I noticed is that for regular queries we require that the name end with `Query` but it looks like we don't for queries generated with `queryName`. Perhaps we need to enforce the same convention with these queries?~

This PR now includes additional validations to prevent this error.

Pull Request resolved: https://github.com/facebook/relay/pull/4663

Reviewed By: evanyeung

Differential Revision: D55728696

Pulled By: captbaritone

fbshipit-source-id: 24b08eb8afd1e0e641c20606c7808d96de894008