Support @catch directive in RelayReader
Summary:
This is initial support for catch - it adds the appropriate errors to the snapshot.

In case of a caught required, it deletes the original required error - and moves it over to a to:RESULT error in fieldErrors.

This does not include providing the error in hooks and such yet.

This is the initial preparation for exposing the result in a catch manner rather than throwing.

Reviewed By: tyao1

Differential Revision: D54806505

fbshipit-source-id: c0b3884fcde1e9f9c7f9ac4e37eef8842588e3d1