Singular weak model type implements interface
Summary:
**Context**

The following (M1-M3) are the milestones we've hit to support abstract types for resolvers:

Note: only strong resolver-defined concrete types are in-scope for M1-M3.

M1: select an interface field on concrete type via spreading the respective fragment on the interface
M2: define a client edge to an interface
M3: define a field directly on an interface via resolvers, such that concrete types do not have to implement the field individually

In M4, we want to add support for additional use cases e.g. **weak types**, concrete types that implement multiple interfaces, and add client schema validation.

**This diff**

Adds tests for functionality added in M1 for weak models, in the case that an interface is implemented by a single weak model.

Reviewed By: captbaritone

Differential Revision: D56311201

fbshipit-source-id: 4167c37e846c3b7a7e62038d104a6c975a63580e