Sort tests for Relay schema generation
Summary:
When adding Relay resolvers for weak types, I ran into an issue where the weak type fields are held in an unorderered data structure. This made the test output non-deterministic as the fields could be printed in any order. To mitigate this, this diff sorts the test output objects before printing them to the snapshot file.

I'm making a separate diff here just to make it clear that tests have not been broken by changes in the next diff.

Reviewed By: tyao1

Differential Revision: D56332103

fbshipit-source-id: 25a54c3fe4d87c9aa347b240acf87a35970e206e