Add weak type definitions to Flow resolvers
Summary:
This diff adds weak objects to Relay resolvers. The overall idea is to take type aliases in Flow and convert them to the `WeakObjectType` docblock IR.

The parser is updated by first adding the types for weak objects and for Flow object types. The parsing is then updated to first look for type aliases inside the relay resolver and create a weak object data structure if one is found. The type alias object is also parsed and turned into a Flow type. These data types are then converted into the IR types in `add_weak_type_definition`.

A number of other small changes were required to the code processing strong object types so that code could be shared between strong and weak object types where possible.

Reviewed By: tyao1

Differential Revision: D56217955

fbshipit-source-id: 6e9ac68066b5ea932bcaad5796ffff3d2e162056