Improve error messages for Relay Resolver schema generation imports
Summary:
This diff adds one new error message and improves another.

The added error message is just to warn not to use namespace imports when importing types for a Relay Resolver. I highly doubt this would come up often (and may error anyways as users would likely have to use the import with a property access) but this makes things a bit clearer and should point users to the best pattern.

The second error message fixes the test added in the previous diff. The error message previously did not contain the actual name of the import value and was mainly concerned with the module and import style. I believe this error message should be much clearer for users.

Reviewed By: captbaritone

Differential Revision: D56382524

fbshipit-source-id: 56ea220388b9101071c16382b13ae11771ee8577