Experimental feature to validate composite schema (#4675)
Summary:
Currently we do not validate the schema. We assume the schema has been pre-validated. This has been fine historically but with Client Schema Extensions and Relay Resolvers, Relay itself is increasingly contributing GraphQL schema in ways that need to be validated.

This change adds experimental support for validating the schema after it's created. It relies on a pre-existing module generally used for standalone schema validation and has its own scheme for error reporting. Specifically, it does not use Diagnostics which are located at a specific position, but rather derives textual error output itself.

This likely won't play well with out LSP. For now I've added it behind a flag without a real error location. This should let us experiment with the feature and explore how we might update the validation to use Diagnostics directly, or at least be able to.

In some cases the server schema may not actually be persisted to disk, in which case the error reporting approach currently used is likely preferable.

This PR also updates all integration tests with invalid schemas to be valid.

Pull Request resolved: https://github.com/facebook/relay/pull/4675

Test Plan:
See D56447664 for an example of enabling this for our tests and the errors it uncovered.

I've also tried enabling this for Ads Manager on WWW (D56450017) which extends the FB schema. It uncovered a number of errors which I've resolved in D56279616 and D56325688. With those fixed it compiles without error.

Perf output of this run can be found here: P1221079983

Looks like for this large schema it took 1671ms. We should look to bring this down before we enable the feature flag. But it does not need to block shipping this diff.

Reviewed By: monicatang

Differential Revision: D56264785

Pulled By: captbaritone

fbshipit-source-id: 93038aefc5bd32f41f8507275c9c2f22d47a6a9f