test existing functionality for a concrete type that implements multiple interfaces
Summary:
**Context**

The following (M1-M3) are the milestones we've hit to support abstract types for resolvers:

Note: only strong resolver-defined concrete types are in-scope for M1-M3.

M1: select an interface field on concrete type via spreading the respective fragment on the interface
M2: define a client edge to an interface
M3: define a field directly on an interface via resolvers, such that concrete types do not have to implement the field individually

In M4, we want to add support for additional use cases e.g. weak types, concrete types that implement *multiple interfaces*, as in this diff, and add client schema validation.

**This diff**

Adds tests for functionality added in M1-M3, for concrete types that implement multiple interfaces.

Reviewed By: captbaritone

Differential Revision: D56161177

fbshipit-source-id: 7c6a560f03db9ceecf5aa1acd4c56af0b61aa715