Add @live to Flow schema generation
Summary:
This diff adds the behavior from the `live` Relay Resolver docblock annotation into the Flow based schema generation. This is indicated by adding the `LiveState` type as the function return.

Note that this does not work on weak objects. It will also not do anything if the `LiveState` type is used wrapped in another Flow type or in the non-return type spot. I think we can make this clear in the documentation, but is there potential for confusion where we should error if LiveState is used in any other position?

Reviewed By: captbaritone

Differential Revision: D56499867

fbshipit-source-id: 5c4df081555c4975ca50c79117b1f4526ab4fe09