Remove package name to remove conflict with npm package
Summary:
This diff is the reponse to a whitehat security report (T187258433). The actual report is in https://www.internalfb.com/intern/whitehat/submission/484430650575877/. Basically, the whitehat found that we have the relay-website package.json defined but not published to npm. By publishing their own package with the same name, they are potentially creating a security threat if we ever depend on the relay-website package. This seems to be a theoretical hack as the security team didn't see any pulls of the package from our ip addresses.

This diff removes the name and version of the relay-website package as it is not published to npm but another package owns the same namespace. According to https://docs.npmjs.com/cli/v10/configuring-npm/package-json#name, the name and version fields are optional for unpublished packages. We don't ever plan on publishing this and just use the package.json for scripts and dependencies. This removes any potential confusion or dependency injection around this package vs the npm "relay-website" package.

Reviewed By: captbaritone

Differential Revision: D56715018

fbshipit-source-id: 9f321d7284044b2a774ac2ae1944e146730145b4