Integrate with Sapling or Git if they exist (#4691)
Summary:
There is some basic code here that is AFAICT only used in Meta's private version of the compiler. We can take that code, spruce it up a bit, and integrate it into the public version of the compiler!

Pull Request resolved: https://github.com/facebook/relay/pull/4691

Test Plan: I built the compiler locally and tested against a local Sapling repo. It worked as expected. I didn't have a chance to try it on a git repo though :/

Reviewed By: captbaritone

Differential Revision: D56963384

Pulled By: bigfootjon

fbshipit-source-id: d9a59ca2a6f3e6d91abea93ff873d966e3492b87