Pass fragment information to Flow schema generation code
Summary: This diff just sets up the next one in the stack by plumbing the graphql fragment information down to the Flow resolver code. It also adds a test to show that using a root fragment in a resolver right now results in no output. I'm just breaking it up to try and keep this "simpler" code out of the next diff and make it easier to review.

Reviewed By: tyao1

Differential Revision: D57033209

fbshipit-source-id: 40b3146078c1bb2c9ebe37bbc109a20f13b75ad9