Add root fragment support to relay resolvers
Summary: This diff allows Flow based resolvers to use fragments as their input, mirroring the "root fragment" in docblock resolvers. It pulls out the relevant features from the graphql string matching the name of the argument to the resolver. It then produces the equivalent docblock IR as if the features were parsed from the docblock.

Reviewed By: tyao1

Differential Revision: D57033324

fbshipit-source-id: c1ae5e4ab0a7e965efd4a595baba5e41b5f3ff64