Sort resolver artifact output
Summary: This diff fixes an issue found in D57033209 where a `Vec` was converted to a `FxHashMap`. This caused the order of the resolvers being processed to be lost. The output of the compiler should be deterministic so I've sorted the results before printing them. It looks like this is a common pattern with other artifacts processed in the file.

Reviewed By: captbaritone

Differential Revision: D57195364

fbshipit-source-id: 62177f43eac188ed4d1d098d48763a3400ad92b9