Update schema generation to use normal Flow AST
Summary:
This diff removes our internal implementation of "FlowType" and replaces it with the base Flow AST received from Hermes parser. This is done as we are slowly needing more and more features from the Hermes Flow AST and are just replicating them in our own FlowType type.

This is a fairly large diff as the changes had to be made pretty atomically. The process I went through was to just replace the `FlowType` type with the `FlowTypeAnnotation` type. Then, anywhere that the `FlowType` fields were used, I inlined the code from the `extract_flow_type` function. This resulted in a bit of code duplication that we can probably clean up into some util functions where it makes things clearer. Overall, the code structure is the same as with the `FlowType` type.

Reviewed By: tyao1

Differential Revision: D57190858

fbshipit-source-id: c6a0edd77f9001b775a3d232d7d7c530cd318dab